+++ e2e/Makefile
@@ -38,8 +38,8 @@ run-test-crud:
 
 .PHONY: run-test-create
 run-test-create:
-	cd .. && npx dotenvx run -f .env.test -- npx playwright test e2e/organization_01_create.test.ts
+	cd .. && npx dotenvx run -f .env.test -- npx playwright test e2e/01_organization_create.test.ts
 
 .PHONY: run-test-list
 run-test-list:
-	cd .. && npx dotenvx run -f .env.test -- npx playwright test e2e/organization_02_list.test.ts
+	cd .. && npx dotenvx run -f .env.test -- npx playwright test e2e/02_organization_list.test.ts
